bugfix: abs_compare expected ints instead of doubles. As a result cut coefficients exceeded 1e8, resulting in numerical instabilities.
bugfix: upper bound tau <= 0 in CGMIP if affine was relaxed in set_mp_bounds() was called.

+++ smips/algorithms/decomposition/aggregator/cgmip/mp_max_coeff.cpp
@@ -1,6 +1,6 @@
 #include "cgmip.h"
 
-static bool abs_compare(int a, int b)
+static bool abs_compare(double a, double b)
 {
   return (std::abs(a) < std::abs(b));
 }
@@ -8,7 +8,7 @@ static bool abs_compare(int a, int b)
 double CGMip::mp_max_coeff()
 {
   double *beta = d_mp.get(GRB_DoubleAttr_X, d_beta.data(), d_beta.size());
-  double beta_max = abs(*max_element(beta , beta + d_beta.size(), abs_compare));
+  double beta_max = abs(*max_element(beta, beta + d_beta.size(), abs_compare));
   delete[] beta;
 
   double alpha = abs(d_alpha.get(GRB_DoubleAttr_X));
